requirements.txt
pandas 
numpy
matplotlib
python-dateutil
mysql
mysql-connector-python
